+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/AuthenticationResponse.kt
@@ -1,28 +1,19 @@
 package at.asitplus.wallet.lib.openid
 
 import at.asitplus.openid.AuthenticationResponseParameters
-import at.asitplus.openid.RelyingPartyMetadata
-import at.asitplus.signum.indispensable.josef.JsonWebKey
 import at.asitplus.wallet.lib.oidvci.OAuth2Error
 
 /**
- * Intermediate DTO to hold the parsed [clientMetadata] as well as the created `params` in [Success].
- *
- * Comes in handy when we need to encrypt the response according to keys passed in [jsonWebKeys].
+ * Intermediate DTO to hold the response to be sent from an [OpenId4VpHolder]
+ * to an [OpenId4VpVerifier] or [DcApiVerifier].
  */
 sealed class AuthenticationResponse(
-    val clientMetadata: RelyingPartyMetadata?,
-    val jsonWebKeys: Collection<JsonWebKey>?,
 ) {
     class Success(
         val params: AuthenticationResponseParameters,
-        clientMetadata: RelyingPartyMetadata?,
-        jsonWebKeys: Collection<JsonWebKey>?,
-    ) : AuthenticationResponse(clientMetadata, jsonWebKeys)
+    ) : AuthenticationResponse()
 
     class Error(
         val error: OAuth2Error,
-        clientMetadata: RelyingPartyMetadata?,
-        jsonWebKeys: Collection<JsonWebKey>?,
-    ) : AuthenticationResponse(clientMetadata, jsonWebKeys)
+    ) : AuthenticationResponse()
 }