+++ vck-openid/src/commonMain/kotlin/at/asitplus/wallet/lib/openid/AuthorizationResponsePreparationState.kt
@@ -6,6 +6,7 @@ import at.asitplus.openid.RequestParametersFrom
 import at.asitplus.openid.VerifierInfo
 import at.asitplus.signum.indispensable.josef.JsonWebKey
 import at.asitplus.wallet.lib.data.CredentialPresentationRequest
+import at.asitplus.wallet.lib.extensions.getEncryptionTargetKey
 import kotlinx.serialization.Serializable
 
 /**
@@ -24,4 +25,9 @@ data class AuthorizationResponsePreparationState(
     /** Whether the request object has been verified (if it was signed at all) */
     val requestObjectVerified: Boolean?,
     val verifierInfo: List<VerifierInfo>?,
-)
+) {
+
+    val responseRequiresEncryption: Boolean
+        get() = request.parameters.responseMode?.requiresEncryption == true
+
+}