+++ CHANGELOG.md
@@ -56,9 +56,10 @@ Release 7.0.0 (unreleased):
 - Digital Credentials API:
     - BREAKING: Remove the `origin` property from Digital Credentials API response models.
  - Deprecations:
-   - Remove code deprecated in 6.0.0, e.g. various `DCAPIWallet*` and related classes, `vckJsonSerializer`
+    - Remove code deprecated in 6.0.0, e.g. various `DCAPIWallet*` and related classes, `vckJsonSerializer`
+    - In `OpenId4VpWallet` deprecate `sendAuthnErrorResponse()` with parameter of type `RequestParametersFrom`, use parameter of type `AuthorizationResponsePreparationState` instead
  - Dependencies:
-   - Update to [Signum 3.24.0](https://github.com/a-sit-plus/signum/releases/tag/3.24.0) for HPKE support
+    - Update to [Signum 3.24.0](https://github.com/a-sit-plus/signum/releases/tag/3.24.0) for HPKE support
 
 Release 6.0.0:
  - JWS: